# api/openapi.yaml
openapi: 3.1.0
info:
  title: Committed API
  version: 0.8.0
  description: |
    HTTP API for Committed — a distributed commit log database.

    Committed exposes two primitives over HTTP:

    - **write**: append proposals to topics (POST /v1/proposal)
    - **configure**: manage databases, ingestables, syncables, and types

    The API is not a query interface, and `/v1/proposal` is write-only:
    there is no endpoint to read the log back. Reads are limited to
    configuration retrieval. For querying committed data, use a syncable to
    replicate it to an external database and query there.

    ## Versioning

    Every API endpoint is served under a `/v1` URL prefix. Additive
    changes (new endpoints, new optional fields) are made in place within
    a major version; a breaking change bumps the prefix to `/v2` and the
    old prefix is supported for at least one further release. Operational
    endpoints (`/health`, `/ready`, `/version`, `/openapi.yaml`, `/docs`)
    are infrastructure, not API surface, and stay unprefixed. See
    `docs/api-compatibility.md` for the full contract.

    ## Authentication

    When the server is started with `COMMITTED_API_TOKEN`, all API
    endpoints require a bearer token in the `Authorization` header. The
    operational endpoints `/health`, `/ready`, `/version`, `/openapi.yaml`,
    and `/docs` are exempt. In development mode (no token configured) no
    authentication is enforced.

    ## Content types

...

    ## Consistency

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/database:
    get:
      tags: [database]
      summary: List all databases
      operationId: getDatabases
      parameters:
        - $ref: '#/components/parameters/Consistency'
      responses:
        '200':
          description: All database configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/database/{id}:
    post:
      tags: [database]
      summary: Create or update a database configuration
      operationId: addDatabase
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        description: Database configuration as TOML or JSON.
        content:
          text/toml:
            schema:
              type: string
          application/json:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/ConfigWritten'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /v1/database/{id}/versions:
    get:
      tags: [database]
      summary: List database configuration versions
      operationId: getDatabaseVersions
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Version history (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/database/{id}/versions/{version}:
    get:
      tags: [database]
      summary: Get a specific database configuration version
      operationId: getDatabaseVersion
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/VersionNumber'
      responses:
        '200':
          description: The configuration at the requested version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/database/{id}/rollback:
    post:
      tags: [database]
      summary: Roll a database back to a previous version
      description: |
        Reads the configuration at version `to` and re-proposes it as a
        new current version. The old versions remain in the history.
      operationId: rollbackDatabase
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/RollbackTo'
      responses:
        '200':
          $ref: '#/components/responses/ConfigWritten'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /v1/ingestable:
    get:
      tags: [ingestable]
      summary: List all ingestables
      operationId: getIngestables
      parameters:
        - $ref: '#/components/parameters/Consistency'
      responses:
        '200':
          description: All ingestable configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/ingestable/{id}:
    post:
      tags: [ingestable]
      summary: Create or update an ingestable configuration
      description: |
        The `[ingestable]` envelope additionally carries the shape-census
        knobs (see `GET /v1/ingestable/{id}/status`): the census is on by
        default; `census = false` opts out, `censusValues = true` opts into
        bounded distinct-value tracking for enum drafting (bounded by
        `censusValueLimit`, default 16).
      operationId: addIngestable
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        description: Ingestable configuration as TOML or JSON.
        content:
          text/toml:
            schema:
              type: string
          application/json:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/ConfigWritten'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
    delete:
      tags: [ingestable]
      summary: Delete an ingestable
      description: |
        Removes an ingestable: its configuration and its checkpoint position are
        deleted atomically (one committed Actual), the worker is stopped, and the
        owner drops the source-side replication resources (for a Postgres source,
        the replication slot and publication) on a best-effort basis. Dropping
        the slot is important on decommission: an orphaned slot holds the source's
        WAL and can fill the source's disk. The logical deletion is the
        authoritative act and completes before this returns; the source teardown
        settles shortly after on the leader and, if it fails, only orphans the
        slot (the delete still succeeds — drop it manually). After deletion a
        later same-id POST starts fresh from a full snapshot. Leader-pinned.
      operationId: deleteIngestable
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Ingestable deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /v1/ingestable/{id}/versions:
    get:
      tags: [ingestable]
      summary: List ingestable configuration versions
      operationId: getIngestableVersions
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Version history (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/ingestable/{id}/versions/{version}:
    get:
      tags: [ingestable]
      summary: Get a specific ingestable configuration version
      operationId: getIngestableVersion
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/VersionNumber'
      responses:
        '200':
          description: The configuration at the requested version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/ingestable/{id}/status:
    get:
      tags: [ingestable]
      summary: Report an ingestable worker's snapshot/streaming progress and lag
      description: |
        Reports an ingestable worker's operational status: snapshot vs.
        streaming phase, per-table snapshot progress, the CDC position (a
        Postgres LSN), how many bytes it is behind the source, and whether it is
        caught up. The ingest analogue of `GET /v1/syncable/{id}/status`. Read
        behind a linearize barrier so the persisted position is consistent on
        whichever node answers; the lag is a live query against the source slot.

        The response also carries the JSON shape `census` the worker took
        during its snapshot pass, per topic: every payload path with its
        observed types, counts, and first/last-seen row (including paths
        nested inside JSON columns); each distinct shape with its row range;
        and a draft JSON Schema (`draftSchema`) ready to review and POST as a
        type. Types and paths only — the draft carries `enum` values only when
        the ingestable opted into `censusValues`. Omitted until a census has
        been published (census opted out, or no snapshot yet).
      operationId: getIngestableStatus
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The ingestable's current worker status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestableStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/type/{id}/pipeline:
    get:
      tags: [type]
      summary: End-to-end status of a topic's pipeline
      description: |
        Composes the end-to-end status of the pipeline flowing through topic
        `{id}` (the type id): its producing ingestable's status (if any), the
        data head, and every syncable consuming the topic, with an overall
        `caughtUp` flag. The linkage (producing ingestable → topic → consuming
        syncables) is resolved server-side from the stored configs — the caller
        passes only the topic. Because it is anchored on the topic rather than an
        ingestable, it also covers a topic fed by direct proposals (no
        ingestable). It reuses the same computations as
        `GET /v1/ingestable/{id}/status` and `GET /v1/syncable/{id}/status`, so
        the numbers match. Read behind a linearize barrier.
      operationId: getPipelineStatus
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The pipeline's composed status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/ingestable/{id}/rollback:
    post:
      tags: [ingestable]
      summary: Roll an ingestable back to a previous version
      operationId: rollbackIngestable
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/RollbackTo'
      responses:
        '200':
          $ref: '#/components/responses/ConfigWritten'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
  /v1/syncable:
    get:
      tags: [syncable]
      summary: List all syncables
      operationId: getSyncables
      parameters:
        - $ref: '#/components/parameters/Consistency'
      responses:
        '200':
          description: All syncable configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
...
  /v1/syncable/{id}/versions:
    get:
      tags: [syncable]
      summary: List syncable configuration versions
      operationId: getSyncableVersions
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Version history (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
  /v1/syncable/{id}/versions/{version}:
    get:
      tags: [syncable]
      summary: Get a specific syncable configuration version
      operationId: getSyncableVersion
      parameters:
        - $ref: '#/components/parameters/Consistency'
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/VersionNumber'
      responses:
        '200':
          description: The configuration at the requested version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
...
  /v1/syncable/{id}/rollback:
    post:
      tags: [syncable]
      summary: Roll a syncable back to a previous version
      operationId: rollbackSyncable
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/RollbackTo'
      responses:
        '200':
          $ref: '#/components/responses/ConfigWritten'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/RequestUnconfirmed'
        '507':
          $ref: '#/components/responses/InsufficientStorage'
...
  /v1/type:
    get:
      tags: [type]
      summary: List all types
      operationId: getTypes
      parameters:
        - $ref: '#/components/parameters/Consistency'
      responses:
        '200':
          description: All type configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/NotLinearizable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token configured via `COMMITTED_API_TOKEN`. If the
        server was started without a token, authentication is disabled
        and any `Authorization` header is ignored.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...