# internal/cluster/db/http/http.go
package http

import (
	"context"
	"net/http"
	"sync"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/rs/cors"
	"go.uber.org/zap"

	"github.com/committeddb/committed/internal/cluster/db"
	"github.com/committeddb/committed/internal/cluster/metrics"
)

// Default CORS request methods and headers, applied when WithCORS is
// given origins but no explicit methods/headers. These cover every verb
// the API actually uses plus the headers a browser client sends on a
// configured request (Content-Type for JSON/TOML bodies, Authorization
// for the bearer token, X-Request-ID for trace correlation).
var (
	defaultCORSMethods = []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"}
	defaultCORSHeaders = []string{"Content-Type", "Authorization", "X-Request-ID"}
)

// defaultMaxBodyBytes is the default request-body cap: 2x the 16 MiB default
// proposal size (db.DefaultMaxProposalBytes), giving headroom for the JSON/TOML
// body being larger than the marshaled proposal it produces. cmd overrides it
// via WithMaxBodyBytes to track a raised proposal cap.
const defaultMaxBodyBytes int64 = 32 * 1024 * 1024

// maxBytes wraps each request body in http.MaxBytesReader so an over-limit body
// fails the read (with *http.MaxBytesError → 413) instead of being buffered into
// memory — an OOM-DoS guard. The propose layer still enforces the exact
// marshaled proposal size; this only bounds what a node will read into RAM.
func maxBytes(limit int64) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			if limit > 0 && r.Body != nil {
				r.Body = http.MaxBytesReader(w, r.Body, limit)
			}
			next.ServeHTTP(w, r)
		})
	}
}

// routePattern returns the matched chi route pattern (e.g. "/v1/proposal") for
// logs and metrics, falling back to the raw path. The pattern is bounded
// cardinality (path params collapse to "{id}"), so it is safe as a metric label.
func routePattern(r *http.Request) string {
	if rc := chi.RouteContext(r.Context()); rc != nil {
		if p := rc.RoutePattern(); p != "" {
			return p
		}
	}
	return r.URL.Path
}

// clusterView is the narrow slice of the engine the transport MIDDLEWARE
// consults — leadership and identity for the leaderRead proxy hop, the
// read-index for the consistency barrier, and the readiness signals. It is
// owned by this package and sized to the middleware (the eudaimonia
// drip.Repository shape): *db.DB satisfies it implicitly, and the in-package
// middleware units stub it for the legs a live single-node engine cannot
// produce (a follower's proxy hop, a failed read-index, an unready node).
// Feature handlers never see it — they hold the engine itself.
type clusterView interface {
	ID() uint64
	Leader() uint64
	AppliedIndex() uint64
	ApplyStalled() bool
	CatchingUp() bool
	LinearizableRead(ctx context.Context) error
	MemberAPIURL(id uint64) (string, bool)
}

type HTTP struct {
	r *chi.Mux
	// db is the engine. Transport lives as a subpackage of the engine
	// precisely so handlers can hold it concretely — no aggregate interface,
	// no engine fakes; tests exercise a real single-node engine through the
	// plugin seams.
	db *db.DB
	// view is the middleware's slice of the engine (see clusterView). In
	// production it IS db; middleware unit tests substitute a stub.
	view             clusterView
	schemas          sync.Map // schemaCacheKey → *jsonschema.Schema
	bearerToken      string   // empty = no auth (dev mode)
	readIndexTimeout time.Duration
	// proxyClient performs the follower→leader hop for leaderRead-wrapped
	// routes (GET /v1/membership). Defaults to a timeout-bounded client with
	// system-root TLS; cmd/node.go overrides it via WithProxyClient to trust
	// the cluster's CA or skip verification for self-signed peer certs.
	proxyClient *http.Client
	// metrics is nil when instrumentation is disabled; call sites nil-check.
	metrics *metrics.Metrics
}

func New(d *db.DB, opts ...Option) *HTTP {
	var o options
	for _, fn := range opts {
		fn(&o)
	}

	r := chi.NewRouter()

	// CORS is off unless an operator supplies an allowlist via
	// WithCORS (COMMITTED_HTTP_CORS_ORIGINS). With no allowlist we mount
	// no CORS middleware at all, so no Access-Control-* headers are
	// emitted and the browser's same-origin policy decides — the right
	// default for a server binary with no in-tree browser client.
	if len(o.corsOrigins) > 0 {
		methods := o.corsMethods
		if len(methods) == 0 {
			methods = defaultCORSMethods
		}
		headers := o.corsHeaders
		if len(headers) == 0 {
			headers = defaultCORSHeaders
		}
		corsMiddleware := cors.New(cors.Options{
			AllowedOrigins: o.corsOrigins,
			AllowedMethods: methods,
			AllowedHeaders: headers,
		})
		r.Use(corsMiddleware.Handler)
	}

	r.Use(securityHeaders)
	r.Use(RequestID)
	// After RequestID so the panic log carries the request ID; before
	// bearerAuth and every handler so the net covers the whole surface.
	r.Use(recoverPanic)

	maxBody := o.maxBodyBytes
	if maxBody <= 0 {
		maxBody = defaultMaxBodyBytes
	}
	// Bound every request body so a large upload can't OOM the node before the
	// propose-layer size check runs. An over-cap body fails the read with 413.
	r.Use(maxBytes(maxBody))

	readIndexTimeout := o.readIndexTimeout
	if readIndexTimeout <= 0 {
		readIndexTimeout = defaultReadIndexTimeout
	}

	proxyClient := o.proxyClient
	if proxyClient == nil {
		proxyClient = &http.Client{Timeout: defaultProxyTimeout}
	}

	h := &HTTP{r: r, db: d, view: d, bearerToken: o.bearerToken, readIndexTimeout: readIndexTimeout, proxyClient: proxyClient, metrics: o.metrics}

	if o.bearerToken != "" {
		zap.L().Info("API bearer-token authentication enabled")
	} else {
		zap.L().Warn("API authentication disabled (no COMMITTED_API_TOKEN set)")
	}

	// Router-level error paths honor the JSON error-envelope contract instead of
	// chi's raw text/plain 404 / empty-body 405. The /v1 subrouter sets its own
	// NotFound below so an unmatched /v1 path is answered from INSIDE the auth
	// group (authenticated) rather than leaking a 404 to an unauthenticated caller.
	r.NotFound(notFoundHandler)
	r.MethodNotAllowed(methodNotAllowedHandler)

	// /health, /ready, /version, /openapi.yaml, and /docs are exempt
	// from authentication — orchestrators need the first two without
	// credentials, /version is useful during rolling upgrades before
	// an operator has a token handy, and the spec + Swagger UI need
	// to be discoverable by new clients before they have a token.
	r.Get("/health", h.Health)
	r.Get("/ready", h.Ready)
	r.Get("/version", h.Version)
	r.Get("/openapi.yaml", h.OpenAPISpec)
	r.Get("/docs", h.SwaggerUI)

	r.Group(func(r chi.Router) {
		if h.bearerToken != "" {
			r.Use(h.bearerAuth)
		}

		// Runtime profiling (/debug/pprof/*), off unless COMMITTED_PPROF enabled
		// it (WithPprof). Mounted inside this group so it inherits bearer auth when
		// a token is configured. middleware.Profiler serves /debug/pprof and the
		// named profiles (heap, goroutine, allocs, profile, trace, ...).
		if o.pprof {
			r.Mount("/debug", middleware.Profiler())
			zap.L().Warn("pprof profiling endpoints enabled at /debug/pprof (COMMITTED_PPROF); disable when not diagnosing")
		}

		// Every API endpoint lives under /v1 so a future breaking change
		// can bump the prefix instead of breaking existing clients. See
		// docs/api-compatibility.md for the versioning contract.
		//
		// The four config resources share a generic handler set (see
		// config_handlers.go); the route table stays explicit so
		// per-resource deviations — type has no rollback and adds
		// migration-errors/migration-retry, syncable adds
		// errors/status/deadletter/replay — remain visible rather than
		// hidden inside a registrar.
		r.Route("/v1", func(r chi.Router) {
			r.Get("/database", h.listConfig("database", h.db.Databases))
			r.Post("/database/{id}", h.addConfig("database", h.db.ProposeDatabase, h.db.DatabaseVersions))
			r.Get("/database/{id}/versions", h.getVersions("database", h.db.DatabaseVersions))
			r.Get("/database/{id}/versions/{version}", h.getVersion("database", h.db.DatabaseVersion))
			r.Post("/database/{id}/rollback", h.rollback("database", h.db.DatabaseVersion, h.db.ProposeDatabase, h.db.DatabaseVersions))

			r.Get("/ingestable", h.listConfig("ingestable", h.db.Ingestables))
			r.Post("/ingestable/{id}", h.addConfig("ingestable", h.db.ProposeIngestable, h.db.IngestableVersions))
			r.Get("/ingestable/{id}/versions", h.getVersions("ingestable", h.db.IngestableVersions))
			r.Get("/ingestable/{id}/versions/{version}", h.getVersion("ingestable", h.db.IngestableVersion))
			r.Get("/ingestable/{id}/status", h.GetIngestableStatus)
			r.Post("/ingestable/{id}/rollback", h.rollback("ingestable", h.db.IngestableVersion, h.db.ProposeIngestable, h.db.IngestableVersions))
			// DELETE is leader-pinned (leaderRead reverse-proxies a follower's
			// request to the leader): the owner-gated source teardown — dropping the
			// Postgres replication slot + publication — runs on the leader, so the
			// request must land where the teardown does.
			r.Delete("/ingestable/{id}", h.leaderRead(h.DeleteIngestable))

			// /proposal is write-only by design: Committed is a commit log,
			// not a query interface. Reads happen on the synced query side
			// (hook up a syncable), not over HTTP. See the API overview in
			// api/openapi.yaml.
			r.Post("/proposal", h.AddProposal)

			// /scrub is the manual right-to-be-forgotten lever: physically
			// remove already-delete-proposed entities from the permanent event
			// log up to the current applied index. Authenticated like every
			// other write; the automatic scheduler does the same on a cadence.
			r.Post("/scrub", h.Scrub)

			// The restatement interpretation registry: append-only statements
			// rebinding how committed actuals are READ (never their bytes).
			// Immutable per id — corrections are new restatements, later wins.
			r.Get("/restatement", h.GetRestatements)
			r.Post("/restatement/{id}", h.AddRestatement)
			r.Post("/restatement/dryrun", h.DryRunRestatement)

			r.Get("/syncable", h.listConfig("syncable", h.db.Syncables))
			r.Post("/syncable/dryrun", h.DryRunSyncable)
			r.Post("/syncable/{id}", h.addConfig("syncable", h.db.ProposeSyncable, h.db.SyncableVersions))
			// DELETE is leader-pinned (leaderRead reverse-proxies a follower's
			// request to the leader): the owner-gated destination teardown runs on
			// the leader and honors the keepData flag recorded there, so the
			// request that carries keepData must land on the same node that tears
			// down.
			r.Delete("/syncable/{id}", h.leaderRead(h.DeleteSyncable))
			r.Get("/syncable/{id}/versions", h.getVersions("syncable", h.db.SyncableVersions))
			r.Get("/syncable/{id}/versions/{version}", h.getVersion("syncable", h.db.SyncableVersion))
			r.Get("/syncable/{id}/errors", h.GetSyncableErrors)
			r.Get("/syncable/{id}/status", h.GetSyncableStatus)
			r.Post("/syncable/{id}/deadletter", h.DeadLetterStuckSyncable)
			// A natural wrong guess the field made: GETting the POST path.
			// Not an alias (aliases breed drift) — a 405 that names where
			// the listing actually lives.
			r.Get("/syncable/{id}/deadletter", func(w http.ResponseWriter, _ *http.Request) {
				writeError(w, http.StatusMethodNotAllowed, "method_not_allowed",
					"dead-letter records are listed at GET /v1/syncable/{id}/errors; POST here skips the proposal a stuck syncable is currently blocked on")
			})
			r.Post("/syncable/{id}/deadletter/{index}/acknowledge", h.AcknowledgeSyncableDeadLetter)
			r.Post("/syncable/{id}/replay/{index}", h.ReplaySyncableDeadLetter)
			r.Post("/syncable/{id}/rollback", h.rollback("syncable", h.db.SyncableVersion, h.db.ProposeSyncable, h.db.SyncableVersions))
			// Rebuild is leader-pinned for the same reason as DELETE: the
			// owner-side destination teardown/re-init runs on the leader.
			r.Post("/syncable/{id}/rebuild", h.syncableOwnerRoute(h.RebuildSyncable))
			r.Post("/syncable/{id}/rematerialize", h.syncableOwnerRoute(h.RematerializeSyncable))

			r.Get("/type", h.listConfig("type", h.db.Types))
			r.Post("/type/{id}", h.addTypeConfig())
			// Deliberate refusal, stated as a posture rather than a bare
			// 405: committed log entries reference their type permanently,
			// so deleting a type would break replay of every entry that
			// carries it. Types are append/version-only; an experiment's
			// leftover type is harmless residue.
			r.Delete("/type/{id}", func(w http.ResponseWriter, _ *http.Request) {
				writeError(w, http.StatusMethodNotAllowed, "types_are_version_only",
					"types cannot be deleted — log entries reference their type permanently, so deletion would break replay; POST a new version to change one (an unused type is harmless residue)")
			})
			r.Get("/type/{id}/versions", h.getVersions("type", h.db.TypeVersions))
			r.Get("/type/{id}/versions/{version}", h.getVersion("type", h.db.TypeVersion))
			r.Get("/type/{id}/migration-errors", h.GetTypeMigrationErrors)
			r.Post("/type/{id}/migration-retry/{index}", h.ReplayTypeMigrationDeadLetter)
			r.Get("/type/{id}/pipeline", h.GetPipelineStatus)

			// Per-node diagnostics. /node/ (not /status) scopes this to the
			// answering node — degraded-build state, applied index, and disk
			// pressure are node-local and ephemeral, unlike the replicated
			// config content. Behind a load balancer you can't target a node,
			// so use it against a node's direct address; /cluster/status below
			// carries the cluster-wide, node-agnostic view.
			r.Get("/node/status", h.NodeStatus)
			// A backup archive of this node, taken live — streamed, so it is
			// scoped to the answering node like /node/status.
			r.Get("/node/backup", h.NodeBackup)

			// Cluster-wide diagnostics that read the same from ANY node (the
			// fan-out sibling reserved by /node/status). Today: the parked-worker
			// summary. Backed by replicated state only, so no leader hop is needed
			// — which is what makes it safe behind a load balancer.
			r.Get("/cluster/status", h.ClusterStatus)

			// Node-to-node: members push their disk state to the leader
			// here and take the cluster write-admission verdict home from
			// the response — the transport for cluster-aware disk
			// admission (see db/disk_cluster.go). Authenticated with the
			// same bearer token as every other write, which is why the
			// token must be cluster-uniform.
			r.Post("/node/disk-report", h.DiskReport)

			// Live cluster membership. GET lists members with their roles
			// and (leader-observed) replication progress; POST adds a voter
			// (or a learner with "learner": true); POST .../promote promotes
			// a learner to a voter; DELETE removes a node — all via
			// joint-consensus (ConfChangeV2) raft reconfiguration.
			// Authenticated like every other write. GET is leaderRead-wrapped
			// so a follower proxies to the leader, giving a load-balanced
			// caller a leader-truthful answer (per-member match index is
			// leader-only state). The node added via POST must first be
			// started in join mode. See docs/operations/membership.md.
			r.Get("/membership", h.leaderRead(h.GetMembership))
			r.Post("/membership", h.AddMember)
			r.Post("/membership/{id}/promote", h.PromoteMember)
			r.Delete("/membership/{id}", h.RemoveMember)

			// Answer unmatched /v1 paths from inside this (authenticated)
			// subrouter, so a probe for a nonexistent /v1 route requires the
			// bearer token and returns the JSON envelope — not chi's
			// unauthenticated default 404 outside the auth group.
			r.NotFound(notFoundHandler)
			r.MethodNotAllowed(methodNotAllowedHandler)
		})
	})

	return h
}

func (h *HTTP) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	h.r.ServeHTTP(w, r)
}